move `exceptions` package into `cli`
