--- .github/workflows/widgetbook-exception.yaml
@@ -1,16 +1,0 @@
-name: widgetbook_exception
-
-on:
-  pull_request:
-    paths:
-      - packages/widgetbook_exception/**
-      - .github/workflows/widgetbook-exception.yaml
-  push:
-    branches:
-      - main
-
-jobs:
-  _:
-    uses: ./.github/workflows/_package.yml
-    with:
-      working_directory: packages/widgetbook_exception
