+++ packages/widgetbook_cli/bin/helpers/exceptions.dart
@@ -1,4 +1,16 @@
-import 'package:widgetbook_exception/widgetbook_exception.dart';
+/// {@template widgetbook_exception}
+/// An exception thrown by an internal widgetbook command.
+/// {@endtemplate}
+abstract class WidgetbookException implements Exception {
+  /// {@macro widgetbook_exception}
+  WidgetbookException(this.message);
+
+  /// The error message which will be displayed to the user .
+  final String message;
+
+  @override
+  String toString() => message;
+}
 
 class WidgetbookDeployException implements WidgetbookException {
   WidgetbookDeployException({String? message})
