+++ packages/widgetbook_cli/pubspec.yaml
@@ -31,8 +31,8 @@ dependencies:
   pubspec: ^2.3.0
   test: ^1.12.1
   version: ^3.0.2
-  widgetbook_exception: ^0.0.1-beta.4
   widgetbook_git: ^2.2.0-beta.9
+
 executables:
   widgetbook: cli_production
   widgetbook_staging: cli_staging
