--- packages/widgetbook_exception/.gitignore
@@ -1,7 +1,0 @@
-# See https://www.dartlang.org/guides/libraries/private-files
-
-# Files and directories created by pub
-.dart_tool/
-.packages
-build/
-pubspec.lock
