--- packages/widgetbook_exception/CHANGELOG.md
@@ -1,16 +1,0 @@
-## 0.0.1-beta.4
-
- - **REFACTOR**: :recycle: removed package analysis_options files.
- - **REFACTOR**: :recycle: warnings.
-
-## 0.0.1-beta.3
-
- - **FEAT**: :sparkles: implement-widgetbook-exception-package. ([3e934f00](https://github.com/widgetbook/widgetbook/commit/3e934f009aa8845be44aa84f84087591bc38c054))
-
-## 0.0.1-beta.2
-
- - **FEAT**: :sparkles: implement-widgetbook-exception-package. ([3e934f00](https://github.com/widgetbook/widgetbook/commit/3e934f009aa8845be44aa84f84087591bc38c054))
-
-## 0.0.1-beta.1
-
-- Initial version.
