--- packages/widgetbook_exception/lib/widgetbook_exception.dart
@@ -1,1 +1,0 @@
-export 'src/widgetbook_exception.dart';
