--- packages/widgetbook_exception/pubspec.yaml
@@ -1,15 +1,0 @@
-name: widgetbook_exception
-description: A general exception for any Widgetbook product.
-version: 0.0.1-beta.4
-homepage: https://www.widgetbook.io?utm_source=pub.dev&utm_medium=link&utm_campaign=widgetbook_exception
-repository: https://github.com/widgetbook/widgetbook
-issue_tracker: https://github.com/widgetbook/widgetbook/issues
-documentation: https://github.com/widgetbook/widgetbook/tree/main/packages/widgetbook-exception#readme
-
-
-environment:
-  sdk: ">=2.18.0 <3.0.0"
-
-dev_dependencies:
-  coverage: ^1.6.2
-  test: ^1.19.2
